refactor(data-view-map): use observable mapApiLinks$ from facade
