--- libs/feature/search/src/lib/metadata-view/data-preview-map/data-preview-map.component.css
@@ -1,5 +1,0 @@
-gn-ui-dropdown-selector {
-  bottom: 0;
-  left: 0;
-  right: 0;
-}
