+++ libs/feature/search/src/lib/metadata-view/data-view-map/data-view-map.component.css
@@ -1,0 +1,5 @@
+gn-ui-dropdown-selector {
+  bottom: 0;
+  left: 0;
+  right: 0;
+}
