+++ libs/feature/search/src/lib/metadata-view/data-view-map/data-view-map.component.html
@@ -1,1 +1,10 @@
-<p>data-view-map works!</p>
+<div class="relative w-full h-full">
+  <gn-ui-map-context [context]="mapContext$ | async"></gn-ui-map-context>
+  <gn-ui-dropdown-selector
+    title="available layers"
+    class="absolute m-2"
+    [choices]="dropdownChoices$ | async"
+    [showTitle]="false"
+    (selectValue)="selectLinkToDisplay($event)"
+  ></gn-ui-dropdown-selector>
+</div>
